small code refactoring to simplify the api. added more documentation to explain implementation details.
