+++ src/firefly/js/util/WebUtil.js
@@ -404,26 +404,6 @@ export function downloadViaAnchor(url, filename) {
     downloadAnchor.click();
 }
 
-export function loadRemoteApi({url, id}) {
-    id = id || 'div_' + new Date().getTime();
-    let apiFrame = document.getElementById(id);
-    if (apiFrame) return;
-
-    apiFrame = document.createElement('iframe');
-    apiFrame.id = id;
-    apiFrame.style.display = 'none';
-    apiFrame.style.width = '0px';
-    apiFrame.style.height = '0px';
-    document.body.appendChild(apiFrame);
-    apiFrame.onload = () => {
-        const myscript = document.createElement('script');
-        myscript.type = 'text/javascript';
-        myscript.src = url;
-        const headEl = apiFrame.contentDocument.getElementsByTagName('head')[0];
-        headEl.appendChild(myscript);
-    };
-}
-
 export function parseUrl(url) {
     const parser = document.createElement('a');
     const pathAry = [];
@@ -806,4 +786,17 @@ function removeEmpties(o) {
             delete o[k]; // The object had no properties, so delete that property
         }
     }
+}
+
+export function hashCode(str) {
+    let hash = 5381;
+    let i = str.length;
+
+    while(i) {
+        hash = (hash * 33) ^ str.charCodeAt(--i);
+    }
+    /* JavaScript does bitwise operations (like XOR, above) on 32-bit signed
+     * integers. Since we want the results to be always positive, convert the
+     * signed int to an unsigned by doing an unsigned bitshift. */
+    return hash >>> 0;
 }
